reuse store behaviors per the WordSpec javadoc
